Move config and util to separated package
