+++ Makefile
@@ -88,4 +88,4 @@ stop-etcd:
 
 .PHONY: test
 test:
-	go test
+	go test -v . ./util
